fix: address critical and high priority CodeRabbit review issues
Fixes 4 issues identified in PR #42 review:

1. CRITICAL: Add user_id to embedding generation in transcript ingestion
   - Pass user_id parameter to generate_embeddings() for cost tracking
   - File: app/services/transcript_service.py:252-255

2. HIGH: Change float to Decimal for monetary fields in ModelPricing
   - Import Decimal from decimal module
   - Update type hints for input_price_per_1m, output_price_per_1m,
     cost_per_request, cache_discount to use Decimal
   - Prevents precision loss and rounding errors
   - File: app/db/models.py:8-9, 368-379

3. HIGH: Add error handling to send_and_save_error helper
   - Wrap DB operations in try-except block
   - Log errors without breaking WebSocket flow
   - File: app/api/websocket/video_loader.py:96-112

4. HIGH: Add error handling to confirmation message persistence
   - Wrap DB operations in try-except block
   - Log errors without failing confirmation flow
   - File: app/api/websocket/video_loader.py:450-467

All changes maintain backward compatibility and improve system robustness.

+++ backend/app/db/models.py
@@ -6,6 +6,7 @@ Uses SQLAlchemy 2.0 declarative mapping style with Mapped and mapped_column.
 """
 
 from datetime import datetime
+from decimal import Decimal
 from typing import List, Optional
 from uuid import UUID, uuid4
 
@@ -364,16 +365,16 @@ class ModelPricing(Base):
     pricing_type: Mapped[str] = mapped_column(
         String(20), nullable=False, comment="per_token, per_request, credit_based"
     )
-    input_price_per_1m: Mapped[Optional[float]] = mapped_column(
+    input_price_per_1m: Mapped[Optional[Decimal]] = mapped_column(
         Numeric(10, 6), nullable=True, comment="For per_token models: cost per 1M input tokens"
     )
-    output_price_per_1m: Mapped[Optional[float]] = mapped_column(
+    output_price_per_1m: Mapped[Optional[Decimal]] = mapped_column(
         Numeric(10, 6), nullable=True, comment="For per_token models: cost per 1M output tokens"
     )
-    cost_per_request: Mapped[Optional[float]] = mapped_column(
+    cost_per_request: Mapped[Optional[Decimal]] = mapped_column(
         Numeric(10, 6), nullable=True, comment="For per_request models: fixed cost per API call"
     )
-    cache_discount: Mapped[Optional[float]] = mapped_column(
+    cache_discount: Mapped[Optional[Decimal]] = mapped_column(
         Numeric(5, 4), nullable=True, comment="Multiplier for cached tokens (e.g., 0.25 for 75% discount)"
     )
     effective_from: Mapped[datetime] = mapped_column(
